\name{Rproj4-package}
\Rdversion{1.1}

\alias{Rproj4-package}

\alias{Rproj4}

\docType{package}
\title{What the package does (short line)}

\description{
More about what it does (maybe more than one line)
}
\author{
Rober J. Hijmans

Maintainer: <r.hijmans@gmail.com>
}

\keyword{ package }
\keyword{ spatial }

